ocr/requirements.txt
boto3>=1.34.0
pdfplumber>=0.10.3
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0
python-pptx>=0.6.21
python-docx>=1.1.0
docx2txt>=0.8